Generate/get or delete an checkout Key

Functions to interact with "checkout keys" on Circle CI.

Check if a specific key type exists in the Circle CI project

create_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

get_checkout_keys(
  repo = github_info()$name,
  user = github_info()$owner$login,
  vcs_type = "gh",
  api_version = "v2"
)

delete_checkout_key(
  fingerprint = NULL,
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "user-key",
  api_version = "v2",
  vcs_type = "gh"
)

has_checkout_key(
  repo = github_info()$name,
  user = github_info()$owner$login,
  type = "github-user-key",
  vcs_type = "gh",
  preferred = "true"
)

Arguments

repo
The Circle CI project. By default the current directory name will be used.

user
The username for the repository. By default queried using `get_user()`.

type
Type of key to add. Options are "github-user-key" and "deploy-key".

api_version
A character string specifying the Circle CI API version. This usually does not need to be changed by the user.

vcs_type
The version control system to use. Defaults to "gh" (Github). Other possible values are "bitbucket".

fingerprint
The fingerprint of the checkout key which should be deleted.

preferred
Checks whether the requested type is the "preferred" key.

Examples

if (FALSE) {
create_checkout_key()

# create only a 'deploy key' pair

create_checkout_key(type = "deploy-key")
}
if (FALSE) {
get_checkout_keys()
}
if (FALSE) {
delete_checkout_key()
}
if (FALSE) {
has_checkout_key()
}