Set up deployment between Circle CI builds and Github repositories

Checks and eventually creates a "user-key" to enable deployment from Circle CI builds.

use_circle_deploy(repo = github_info()$name, user = github_info()$owner$login)

Arguments

repo
The Circle CI project. By default the current directory name will be used.

user
The username for the repository. By default queried using `get_user()`.

Details

The easiest way to achieve a deployment from Circle CI builds to a Github repo is by creating a so called "user-key" with your Circle CI account. Essentially, this is a SSH key pair.

This function checks for the presence of such and, if missing, creates one.

If the repo has not been enabled yet on Circle CI, please run `enable_repo()` first. Also to be able to log in to Github, you will need to have a `GITHUB_TOKEN` being set as an environment variable. If you are unsure if you have done this already, call `usethis::github_token()`. If none is set, this function will prompt you to create one.

Examples

if (FALSE) {
use_circle_deploy()
}