Authenticate to Circle CI

A circle API Key is needed to interact with the Circle CI API. `browse_circle_token()` opens a browser window for the respective circle endpoint. On this site, you can copy your personal API key and then follow the instructions of the console output or the ones shown below.

browse_circle_token()

Store API Key

The `circle` package supports two ways of storing the Circle API key(s):

- via env vars `R_CIRCLE` - via `~/.circleci/cli.yml`

The latter should already be present if you already used the `circle` CLI tool at some point in the past. If not, its up to your preference which approach to use.

The following instructions should help to set up `~/.circleci/cli.yml` correctly: 1. Copy the token from the browser window which just opened. You can use `edit_circle_config()` to open `~/.circleci/cli.yml`. 2. The token should be stored using the following structure

``` host: https://circleci.com endpoint: graphql-unstable token: <token> ```

Examples

if (FALSE) {
browse_github_token()

edit_circle_config()
}